Add a shell script to build and boot Linux
Shell script will download prompt the user to download and build the
bootloader and device tree required to boot linux on the Sail model,
then download a Kernel if propmted before building it with a known
good .configure file.

Currently we expect that CROSS_COMPILE can be set to
aarch64-linux-gnu-, but this won't be true on all platforms so we
should make this configurable.