Add 32-bit instructions
Split the model into various files, since it's so big:

- aarch_types.sail contains simple type definitions
- aarch_mem.sail for various memory related functions
- aarch64.sail contains all non-vector and non-float 64-bit instructions
- aarch64_float.sail constains aarch64 floating point
- aarch64_vector.sail contains aarch64 vector instructions
- aarch32.sail contains all 32-bit instructions
- aarch_decode.sail has all the decodes

main.sail has been updated to follow the structure of elfmain.sail,
although there are still some updates to both for proper 32-bit
architecture support.

Unfortunately everything must be included to build the full model, as
the aarch_decode file currently pulls in everything. This should be
changed in future - the build times are really painful right now, as
this is far larger than anything we've used Sail for previously, so we
should look into that.