Fix some ASL bugs
In AlignmentFault fault.level is left uninitialized, but the ASL really depends on
the fact that it is zero (or at the very least that it is not -1

The fault.gpcf.gpf check in AArch64_Abort is also missing a HaveRME() guard. This
is a similar issue, where the gpcf field is left uninitialized in the RME disabled
case, but the ASL is relying on the fact that the unitialized value just happens to
be GPCF_None due to its position in the enumeration. Isla currently catches this so
we add the guard back in.