(armv9p4) TEMP: fix: missing stub for NV
on an `MRS Xn, CurrentEL` if FEAT_NV is implemented and execution
is currently at EL1, and nested virtualisation is enabled,
the read should return 2 instead.

The stub that actually does the check is missing,
so here we 'fix' it by hardcoding false
i.e. that FEAT_NV is not implemented.