Move the PC write to the earliest point where next instruction is known
This is after any possible exceptions (in particular translation) and
after branching choices if the current instruction is a branch.  To keep
PC reads after this point the same, we use the original `_PC` register
as a global intra-instruction variable, and add a new `PC` register.

In supported instructions:
* For memory instruction, just after translation (multiple data accesses
  in a single instruction are not supported)
* For branching instruction, when we know direction in `BranchTo` or
  `BranchToAddr`
* For all other instruction that can't trigger an exception, just after
  decode

Unsupported instructions (including vector, FP, PAC, and MTE) don't set
`PCUpdated` and will be detected near the end of `__InstructionExecute`.